Pre pruebas unitarias

+++ Banco_Vejarano/src/Cuenta.java
@@ -1,7 +1,5 @@
 public class Cuenta extends Producto {
-    public Cuenta(Cliente cli, int numero, String tipo, String estado, int saldo) {
+    public Cuenta(Cliente cli, int numero, String tipo, String estado, double saldo) {
         super(cli, numero, tipo, estado, saldo);
     }
-
-
 }
