+++ Banco_Vejarano/src/Prestamo.java
@@ -1,5 +1,5 @@
 public class Prestamo extends Producto {
-    public Prestamo(Cliente cli, int numero, String tipo, String estado, int saldo) {
+    public Prestamo(Cliente cli, int numero, String tipo, String estado, double saldo) {
         super(cli, numero, tipo, estado, saldo);
     }
 }
